# Concurrently run BM25 search terms