### What's being changed:
Run each query term in parallel in BM25

new
```
Query Latencies
--------------------
Min:    3.210625ms
Mean: 580.462948ms
Max:  1.730826167s
p50:  672.474375ms
p90:  1.017810542s
p99:  1.730826167s
```
old:
```
Query Latencies
--------------------
Min:    2.413708ms
Mean: 809.687782ms
Max:  3.950294042s
p50:  752.353667ms
p90:     1.701873s
p99:  3.950294042s

```

### Review checklist

- [x] All new code is covered by tests where it is reasonable.
- [x] Performance tests have been run or not necessary.